+++ libs/langgraph-checkpoint-mongodb/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ---
 
+## Changes in version 0.2.0 (TBD)
+
+- Implements async methods of MongoDBSaver.
+- Deprecates ASyncMongoDBSaver, to be removed in 0.3.0
+
 ## Changes in version 0.1.4 (2025/06/13)
 
 - Add TTL (time-to-live) indexes for automatic deletion of old checkpoints and writes